requirements.txt
Flask>=2.2.2
PyYAML>=3.12